+++ GameProject/GameProject/ObjectContainer.cpp
@@ -7,7 +7,7 @@ void ObjectContainer::load(sf::VideoMode res)
 	const int max_objects = 4;
 	for (int i = 0; i < max_objects; i++)
 	{
-		v.push_back(Object(res, 60, i * (res.width / max_objects)));
+		v.push_back(Object(res, 60, i * (float)(res.width / max_objects)));
 	}
 }
